requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-bootstrap3==9.1.0
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-suit==0.2.25
django-tables2==1.19.0
django-widget-tweaks==1.4.1
ephem==3.7.6.0
idna==2.6
kombu==4.1.0
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
requests==2.18.4
urllib3==1.22
vine==1.1.4
